refactor(server): register explicit vercel api routes
